# perf(cli|main): lazy loads config mangling, extract and enrich steps