--- a/configs/.dependency-cruiser-show-metrics-config.mjs
+++ b/configs/.dependency-cruiser-show-metrics-config.mjs
@@ -131,6 +131,10 @@ export default {
           ],
           dependencies: [
             {
+              criteria: { dynamic: true },
+              attributes: { style: "dashed" },
+            },
+            {
               criteria: { "rules[0].severity": "error" },
               attributes: { fontcolor: "red", color: "red" },
             },