## Description

- in the cli lazy loads all configuration file pre-processing
- in the main `cruise` routine does the same for the `extract` and `enrich` steps

## Motivation and Context

It shaves off some time when running dependency-cruiser from the cache. I had hoped to shave off hundreds of milliseconds (so loading e.g. typescript, babel etc was postponed), but that's going to take a larger re-factoring.

## How Has This Been Tested?

- [x] green ci
- [x] performance tests (see below for results)

## Comparison

For runs that are served for 100% out of the cache the win is about 47ms total run time (measured on the same environment and conditions as #795). For regular runs, without cache, the difference is negligable: 3ms slower on the benchmark run, which falls within the calculated error margin (which is 10-12ms, from the runs below).

### Before ('develop' branch)

#### cached
```
$ hyperfine --warmup 3 --runs 30 "bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --cache node_modules/.cache/dependency-cruiser/load-cached --no-progress"
Benchmark 1: bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --cache node_modules/.cache/dependency-cruiser/load-cached --no-progress
  Time (mean ± σ):      1.283 s ±  0.006 s    [User: 1.343 s, System: 0.204 s]
```

#### regular run
```
$ hyperfine --warmup 3 --runs 30 "bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress"
Benchmark 1: bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress
  Time (mean ± σ):      2.984 s ±  0.012 s    [User: 4.626 s, System: 0.295 s]
  Range (min … max):    2.963 s …  3.010 s    30 runs
```

### After (this branch)

#### cached
```
$ hyperfine --warmup 3 --runs 30 "bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --cache node_modules/.cache/dependency-cruiser/load-cached --no-progress"
Benchmark 1: bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --cache node_modules/.cache/dependency-cruiser/load-cached --no-progress
  Time (mean ± σ):      1.236 s ±  0.004 s    [User: 1.291 s, System: 0.194 s]
  Range (min … max):    1.226 s …  1.246 s    30 runs
```

#### regular run
```
$ hyperfine --warmup 3 --runs 30 "bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress"
Benchmark 1: bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress
  Time (mean ± σ):      2.987 s ±  0.009 s    [User: 4.658 s, System: 0.294 s]
  Range (min … max):    2.971 s …  3.008 s    30 runs
```

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [x] Refactor (non-breaking change which fixes an issue without changing functionality)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

- [x] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [x] :balance_scale:
  - The contribution will be subject to [The MIT license](https://github.com/sverweij/dependency-cruiser/blob/develop/LICENSE), and I'm OK with that.
  - The contribution is my own original work.
  - I am ok with the stuff in [**CONTRIBUTING.md**](https://github.com/sverweij/dependency-cruiser/blob/develop/.github/CONTRIBUTING.md).